# docs/swagger.yaml
basePath: /
definitions:
...
  database.CreateProjectParams:
    properties:
      data:
        additionalProperties:
          type: string
        example:
          key1: value1
          key2: value2
        type: object
      name:
        example: Sample Project
        type: string
    required:
    - name
    type: object
...
  database.Project:
    properties:
      created_at:
        example: "2021-12-01 13:00:00"
        type: string
      data:
        additionalProperties:
          type: string
        example:
          key1: value1
          key2: value2
        type: object
      id:
        example: "1234567812345678"
        type: string
      name:
        type: string
      updated_at:
        example: "2021-12-01 13:00:00"
        type: string
    type: object
...
  util.MultiError:
    properties:
      errors:
        example:
        - Error message1
        - Error message2
        items:
          type: string
        type: array
    type: object
  util.NotFound:
    properties:
      error:
        example: Not Found
        type: string
    type: object
  util.Schedule:
    properties:
      arrival:
        type: string
      breakID:
        type: integer
      createdAt:
        type: string
      departure:
        type: string
      endLoad:
        items:
          type: integer
        type: array
      id:
        type: integer
      jobID:
        type: integer
      projectID:
        type: integer
      serviceTime:
        type: integer
      shipmentID:
        type: integer
      startLoad:
        items:
          type: integer
        type: array
      travelTime:
        type: integer
      type:
        type: integer
      updatedAt:
        type: string
      vehicleID:
        type: integer
      waitingTime:
        type: integer
    type: object
  util.Success:
    properties:
      success:
        example: "true"
        type: string
    type: object
host: localhost:9100
info:
  contact:
    email: info@georepublic.de
    name: Team Georepublic
  description: This is an API for scheduling VRP tasks. Source code can be found on
    https://github.com/Georepublic/pg_scheduleserv
  license:
    name: GNU Affero General Public License
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: https://swagger.io/terms/
  title: pg_scheduleserv API
  version: 0.1.0
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: Get a list of projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: List projects
      tags:
      - Project
    post:
      consumes:
      - application/json
      description: Create a new project with the input payload
      parameters:
      - description: Create project
        in: body
        name: Project
        required: true
        schema:
          $ref: '#/definitions/database.CreateProjectParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Create a new project
      tags:
      - Project
  /projects/{project_id}:
    delete:
      consumes:
      - application/json
      description: Delete a project with its project_id
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Delete a project
      tags:
      - Project
    get:
      consumes:
      - application/json
      description: Fetch a project with its project_id
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.NotFound'
      summary: Fetch a project
      tags:
      - Project
    patch:
      consumes:
      - application/json
      description: Update a project with its project_id
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Update project
        in: body
        name: Project
        required: true
        schema:
          $ref: '#/definitions/database.CreateProjectParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Update a project
      tags:
      - Project
...
  /projects/{project_id}/schedule:
    delete:
      consumes:
      - application/json
      description: Delete the schedule for a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Delete the schedule
      tags:
      - Schedule
    get:
      consumes:
      - application/json
      description: Get the schedule for a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - text/calendar
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Schedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Get the schedule
      tags:
      - Schedule
    post:
      consumes:
      - application/json
      description: Schedule the tasks present in a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Schedule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.MultiError'
      summary: Schedule the tasks
      tags:
      - Schedule
...
schemes:
- http
- https
swagger: "2.0"